Added gene expression URL updates to the model_overview transform and fixed a few model_details tests
