+++ tests/test_assets/model_overview/input/model_overview_model_info_extra_column_input.csv
@@ -1,4 +1,4 @@
-name,matched_controls,model_type,contributing_group,study_synid,rrid,jax_id,alzforum_id,genotype,aliases,extra_column
-3xTg-AD,B6129,Familial AD,UCI,syn22964685,MMRRC_034830-JAX,004807,3xtg,"B6;129-Tg(APPSwe,tauP301L)1Lfa <i>Psen1<sup>tm1Mpm</sup></i>/Mmjax",3xTg,extra_value
-5xFAD (UCI),C57BL6J,Familial AD,UCI,syn16798076,MMRRC_034848-JAX,008730,5xfad-c57bl6,"B6.Cg-Tg(APPSwFlLon,PSEN1*M146L*L286V)6799Vas/Mmjax","5XFAD, 5x-FAD, Tg6799",extra_value
-APOE4,C57BL6J,Late Onset AD,IU/JAX/PITT,syn17095980,IMSR_JAX:027894,027894,apoe4-knock-jax,B6(SJL)-<i>Apoe<sup>tm1.1(APOE*4)Adiuj</sup></i>/J,"APOE4 KI, APOE*4 KI",extra_value 
+name,matched_controls,model_type,contributing_group,study_synid,rrid,jax_id,alzforum_id,genotype,aliases,url_categories_value,url_models_value,extra_column
+3xTg-AD,B6129,Familial AD,UCI,syn22964685,MMRRC_034830-JAX,004807,3xtg,"B6;129-Tg(APPSwe,tauP301L)1Lfa <i>Psen1<sup>tm1Mpm</sup></i>/Mmjax",3xTg,,,extra_value
+5xFAD (UCI),C57BL6J,Familial AD,UCI,syn16798076,MMRRC_034848-JAX,008730,5xfad-c57bl6,"B6.Cg-Tg(APPSwFlLon,PSEN1*M146L*L286V)6799Vas/Mmjax","5XFAD, 5x-FAD, Tg6799",,,extra_value
+APOE4,C57BL6J,Late Onset AD,IU/JAX/PITT,syn17095980,IMSR_JAX:027894,027894,apoe4-knock-jax,B6(SJL)-<i>Apoe<sup>tm1.1(APOE*4)Adiuj</sup></i>/J,"APOE4 KI, APOE*4 KI",,,extra_value
