+++ tests/test_assets/model_overview/input/model_overview_model_info_url_test_good_input.csv
@@ -1,0 +1,5 @@
+name,matched_controls,model_type,contributing_group,study_synid,rrid,jax_id,alzforum_id,genotype,aliases,url_categories_value,url_models_value
+Model_url_default,C57BL/6J,Tauopathy,IU/JAX/PITT,syn25539881,IMSR_JAX:005491,5491,htau,B6.Cg-Mapttm1(EGFP)Klt Tg(MAPT)8cPdav/J,"hTau, Mapt-/-",,
+Model_url_categories,B6129,Familial AD,UCI,syn22964685,MMRRC_034830-JAX,4807,3xtg,"B6;129-Tg(APPSwe,tauP301L)1Lfa <i>Psen1<sup>tm1Mpm</sup></i>/Mmjax",3xTg,category_string,
+Model_url_models,C57BL/6J,Tauopathy,IU/JAX/PITT,syn25539881,IMSR_JAX:005491,5491,htau,B6.Cg-Mapttm1(EGFP)Klt Tg(MAPT)8cPdav/J,"hTau, Mapt-/-",,"model1,model2"
+Model_url_both,B6129,Familial AD,UCI,syn22964685,MMRRC_034830-JAX,4807,3xtg,"B6;129-Tg(APPSwe,tauP301L)1Lfa <i>Psen1<sup>tm1Mpm</sup></i>/Mmjax",3xTg,category_string,"model3,model4"
