+++ tests/test_assets/model_overview/input/model_overview_model_results_info_url_test_input.csv
@@ -1,0 +1,5 @@
+name,gene_expression,disease_correlation,pathology,biomarkers
+Model_url_default,TRUE,,,
+Model_url_categories,TRUE,,,
+Model_url_models,TRUE,,,
+Model_url_both,TRUE,,,
